Begun work on importing excel files
I've begun cleaning up a bit by removing the table, and adding a dropbox instead.  The drag and drop doesn't work yet, BUT if you click on it you can select your file.  The contents of the file will then be outputted to the console log.  In the future we can split up the data and send it to the  proper place.