Front end lookup table (#68)
* Begun work on importing excel files

I've begun cleaning up a bit by removing the table, and adding a dropbox instead.  The drag and drop doesn't work yet, BUT if you click on it you can select your file.  The contents of the file will then be outputted to the console log.  In the future we can split up the data and send it to the  proper place.

* Code Cleanup and Some Documentation

Cleaned up some imports, added some newlines to the end of files, added some comments to the files to document functionality.

* Edited Backend-Drop-Sim_schem

Getting ready to connect to backend.

* End-to-end functionality, with help from Brandon

* About Page

Did more with the about page.

Co-authored-by: Brandon Miller <8723104+ScratchnSniff0@users.noreply.github.com>